FastAPI/blog/token.py
from jose import JWTError, jwt
from datetime import datetime, timedelta


SECRET_KEY = 'de9e8fb5af98de8fce4fe16e9d22e7fe9fce195edab68ddeb0d6bbd1a668da6f'
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt